refactor(docker-tests): extract UTXO swap tests to dedicated module
Extract 26 UTXO-only swap tests from docker_tests_inner.rs to new
utxo_swaps_v1_tests.rs module, gated by docker-tests-swaps-utxo feature.

Tests extracted:
- Swap spend/refund mechanics (test_search_for_swap_tx_spend_*)
- Max volume tests (test_get_max_taker_vol*, test_get_max_maker_vol*)
- UTXO merge/consolidation (test_utxo_merge*, test_consolidate_utxos_rpc)
- Locked coins tests (test_buy/sell_when_coins_locked_by_other_swap)
- Order type tests (test_fill_or_kill_*, test_gtc_*)
- Trade tests (test_trade_base_rel_mycoin_mycoin1_*)

This enables running UTXO swap tests independently without ETH/ERC20
containers, reducing CI feedback time for UTXO-only changes.
